vex fix
